+++ docs/history.md
@@ -19,6 +19,7 @@
   - Blisters Baby (#240) - Shot speed oscillates
   - Cupcake Baby (#321) - Starts with Voodoo Head + Flat File
   - Dark Elf Baby (#378) - Starts with Bone Spurs + Polished Bone
+  - Puncher Baby (#416) - Spawns a Golden Troll Bomb in every room
   - Tired Baby (#443) - Starts with 10x Cube Baby + Ice Cube
   - Burning Baby (#467) - Starts with Vasculitis + A Lighter + 40 Luck
   - Tooth Baby (#478) - Starts with Black Tooth